More conversion methods. More organization.
